+++ src/index.ts
@@ -1,2 +1,3 @@
 export { default as plugin } from "./plugin.ts";
+export { default as strictConfig } from "./strict-config.ts";
 export { strictRules } from "./strict-rules.ts";